refactor: move all scaling processors to processing.scaling
