+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/processing/scaling/ScaledUpApplier.java
@@ -5,14 +5,22 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.engine.scaling;
+package io.camunda.zeebe.engine.processing.scaling;
 
 import io.camunda.zeebe.engine.state.TypedEventApplier;
+import io.camunda.zeebe.engine.state.mutable.MutableRoutingState;
 import io.camunda.zeebe.protocol.impl.record.value.scaling.ScaleRecord;
 import io.camunda.zeebe.protocol.record.intent.scaling.ScaleIntent;
 
-public class ScaleUpStatusResponseApplier implements TypedEventApplier<ScaleIntent, ScaleRecord> {
+public class ScaledUpApplier implements TypedEventApplier<ScaleIntent, ScaleRecord> {
+  private final MutableRoutingState routingState;
+
+  public ScaledUpApplier(final MutableRoutingState routingState) {
+    this.routingState = routingState;
+  }
 
   @Override
-  public void applyState(final long key, final ScaleRecord value) {}
+  public void applyState(final long key, final ScaleRecord value) {
+    // TODO: when relocation is done, apply the changes to the routing state
+  }
 }
