+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/processing/scaling/ScalingProcessors.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.engine.scaling;
+package io.camunda.zeebe.engine.processing.scaling;
 
 import io.camunda.zeebe.engine.processing.distribution.CommandDistributionBehavior;
 import io.camunda.zeebe.engine.processing.streamprocessor.TypedRecordProcessors;
