+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/state/appliers/EventAppliers.java
@@ -7,9 +7,9 @@
  */
 package io.camunda.zeebe.engine.state.appliers;
 
-import io.camunda.zeebe.engine.scaling.ScaleUpStatusResponseApplier;
-import io.camunda.zeebe.engine.scaling.ScaledUpApplier;
-import io.camunda.zeebe.engine.scaling.ScalingUpApplier;
+import io.camunda.zeebe.engine.processing.scaling.ScaleUpStatusResponseApplier;
+import io.camunda.zeebe.engine.processing.scaling.ScaledUpApplier;
+import io.camunda.zeebe.engine.processing.scaling.ScalingUpApplier;
 import io.camunda.zeebe.engine.state.EventApplier;
 import io.camunda.zeebe.engine.state.EventApplier.NoSuchEventApplier.NoApplierForIntent;
 import io.camunda.zeebe.engine.state.EventApplier.NoSuchEventApplier.NoApplierForVersion;
